+++ CHANGELOG.md
@@ -10,8 +10,21 @@ In this release we have mostly updated our dependencies, but we also found a bug
 
 - Update dependencies
 - Improved sentry support
-- Sort historical data for `TemplateLines`
+- Sort historical data for `templateLines`
 - Refactoring graph helpers
+- Refactoring graph templates (4 of 12)
+  - ✅ `templateBars`
+  - ✅ `templateBoxPlot`
+  - ✅ `templateBarsAndLines`
+  - ✅ `templateAgeDistribution`
+  - ❌ `templateLines`
+  - ❌ `templateLinesMulti`
+  - ❌ `templateLivingConditions`
+  - ❌ `templatePopulationDetailsTable`
+  - ❌ `templatePyramid`
+  - ❌ `templateStackedBars`
+  - ❌ `templateTernaryPlot`
+  - ❌ `templateComboHistogram`
 
 # Release v1.1.1